requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
psycopg2-binary==2.9.6
apache-airflow==2.6.3